+++ .idea/.name
@@ -1,1 +1,1 @@
-Pytesseract
+main.py
